requirements.txt
streamlit==1.36.0
google-generativeai==0.7.1
python-dotenv==1.0.1
Pillow==10.4.0 # For image manipulation